+++ spotbugsTestCases/src/java/privateFinalLocks/UnsafeMethodSynchronizationWithClassExposingItSelf.java
@@ -1,6 +1,6 @@
 package privateFinalLocks;
 
-class BadMethodSynchronizationWithClassExposingItSelf {
+class UnsafeMethodSynchronizationWithClassExposingItSelf {
     public synchronized void doStuff() { // Locks on the object's monitor(intrinsic lock), bug should be detected here
         System.out.println("Do some stuff");
     }
@@ -9,11 +9,11 @@ class BadMethodSynchronizationWithClassExposingItSelf {
         System.out.println("Do some stuff");
     }
 
-    public static BadMethodSynchronizationWithClassExposingItSelf lookup(String name) {
+    public static UnsafeMethodSynchronizationWithClassExposingItSelf lookup(String name) {
         return null;
     }
 
-    public static BadMethodSynchronizationWithClassExposingItSelf lookup2(String name) {
+    public static UnsafeMethodSynchronizationWithClassExposingItSelf lookup2(String name) {
         return null;
     }
 }