+++ spotbugsTestCases/src/java/privateFinalLocks/UnsafeMethodSynchronizationWithPublicStaticSynchronization.java
@@ -1,6 +1,6 @@
 package privateFinalLocks;
 
-public class BadMethodSynchronizationWithPublicStaticSynchronization {
+public class UnsafeMethodSynchronizationWithPublicStaticSynchronization {
     public static synchronized void doStuff() { // locking on the class, bug should be detected here
         System.out.println("Do some stuff");
     }